pull out relationship loading code
